Added comprehensive example about node parameters
@fabrizioromanelli in the last two days I throughly explored the potential of the parameters system.
The result is here. I think this can help us with a lot of things without hurting efficiency, if used
correctly. We can define or update parameters at startup, runtime, parse configuration files, declare them
in launch files and even enforce internal node updates and operations when they are changed.
Example code for every aspect is included in this package.