feat: add actions examples
@fabrizioromanelli the action system is very complex, and can be customized in many ways.
Unfortunately the official documentation doesn't cover it at all, and is full of errors.
This package, that generates 4 executables, contains ready-to-use code and covers almost every aspect.
What I left out is not so necessary or can be easily inferred by the rest.
All source files are full of comments that explain how it all works: there's too many features to even
list here, but the code should be comprehensive and self-explanatory.