feat: add custom context example
@fabrizioromanelli this is a small feature, but in some situations (e.g. supervisor/FSM)
it may be critical: the ability to specify custom cleanup tasks to be executed
upon process termination (invoked by a signal). When we initialize the middleware by calling
rclcpp::init it initializes the "global context" (aka "DDS Participant" in DDS Wire Protocol
terminology, DDS instance+ROS stuff in layman's terms), but it also registers custom
signal handlers for common termination signals that don't do much more than a log print.
Also, if you want to manually specify when the ROS 2 context has to be shut down
during cleanup you can't do that; both things led to all sorts of deadlocks and the like
last year when apps had to be terminated asynchronously (to e.g. handle a fault).
This little example shows which configurations must be made to customize everything.