feat: actions_example_cpp
@lorenzo-bianchi @giorgiomanca97 @alessandrotenaglia @smatto72
Actions API changes in Humble:
- client: goal_response_callback takes now a single argument of type ACTIONGoalHandleSharedPtr (they were two before). This reduces the number of futures hanging around (and in case one needs futures, shared_future objects may be created with new methods of these objects).

@lorenzo-bianchi @giorgiomanca97
We should consider adding a package to dua-utils which wraps synchronous and asynchronous action calls, similarly to what we did in Python, and also standardize that as a Python package as well. In C++ we need templates. We should also add methods to the DUANode base class to instantiate these clients. Let's discuss this tomorrow.